Add 'changed by' methods
Why:

Changes from 100 to 200 were being represented as a 200% increase,
instead of a less-misleading 100% increase.

We changed it so the methods `a.change.of` and `changed.by` return the
full difference -- in this case, the 200% -- whereas the increase and
decrease take into account the starting conditions.

Some helpers are still a little misleading, and should be addressed.
They're noted in the README.

We also add a `configure` block, which allows the user to set
significant digits, and/or whether to convert all numbers to integers.